Docker - Face analysis service/requirements.txt
fastapi
uvicorn
scipy
Pillow
numpy